Refactor Distributore and DistributoreService; rename status to stato and update data types for latitude and longitude

+++ Progetto-Coffee-cApp/src/main/java/it/unipa/wsda/progettocoffeecapp/model/Distributore.java
@@ -16,8 +16,8 @@ public class Distributore {
     @Enumerated(EnumType.STRING)
     private StatiDistributori stato;
 
-    private Integer lat;
-    private Integer lon;
+    private Double lat;
+    private Double lon;
     private String edificio;
     private String piano;
